// plugin/dev/src/endpoints/seed/post-1.d.ts
import type { Media, User } from '@/payload-types';
import type { RequiredDataFromCollectionSlug } from 'payload';
export type PostArgs = {
    author: User;
    blockImage: Media;
    heroImage: Media;
};
export declare const post1: ({ author, blockImage, heroImage }: PostArgs) => RequiredDataFromCollectionSlug<"posts">;